# Conflicts:
#	app/src/main/java/com/example/danie/flexicuapplication/GUI/RentOut.java

+++ app/src/main/java/com/example/danie/flexicuapplication/LogicLayer/CriteriaPay.java
@@ -6,8 +6,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class CriteriaPay implements CriteriaInterface {
-    private int searchCriteria;
-    public CriteriaPay(int searchCriteria)
+    private double searchCriteria;
+    public CriteriaPay(double searchCriteria)
     {
         this.searchCriteria = searchCriteria;
     }
